+++ GrammaticalEvolution/src/main/java/model/module/operator/collector/FitnessCollectorOperator.java
@@ -31,7 +31,7 @@ public class FitnessCollectorOperator extends CollectorOperator{
 	@Override
 	public void collect() {
 		DoubleSummaryStatistics stats = this.objetivePopulation.stream().filter(Individual::isValid).mapToDouble(Individual::getFitness).summaryStatistics();
-		Individual best = this.objetivePopulation.stream().max((e1,e2)->Float.compare(e1.getFitness(), e2.getFitness())).get();
+		Individual best = this.objetivePopulation.stream().max((e1,e2)->Double.compare(e1.getFitness(), e2.getFitness())).get();
 		//if(best.isValid())System.out.println(best.getPhenotype().getVisualCode());
 		if(best.isValid())System.out.println(best.getPhenotype().getPlainSymbols());
 		System.out.println("Best Individual: "+stats.getMax());
